requirements.txt
python-dotenv == 1.0.0
pyfiglet == 1.0.2
google-api-core == 2.11.0
google-api-python-client == 2.88.0
google-auth == 2.19.0
google-auth-httplib2 == 0.1.0
google-auth-oauthlib == 1.0.0
googleapis-common-protos == 1.61.0
